backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.35
google-generativeai>=0.3.1
typing-extensions>=4.8.0
python-multipart==0.0.6